+++ src/device-types/HANumber.cpp
@@ -42,9 +42,9 @@ void HANumber::buildSerializer()
         return;
     }
 
-    _serializer = new HASerializer(this, 15); // 15 - max properties nb
-
+    _serializer = new HASerializer(this, 16); // 16 - max properties nb
     _serializer->set(AHATOFSTR(HANameProperty), _name);
+    _serializer->set(AHATOFSTR(HAObjectIdProperty), _objectId);
     _serializer->set(HASerializer::WithUniqueId);
     _serializer->set(AHATOFSTR(HADeviceClassProperty), _class);
     _serializer->set(AHATOFSTR(HAIconProperty), _icon);